// frontend/src/types/concert.ts
export interface Concert extends BaseEntity {
  title: string;
  description: string;
  location: string;
  locationX: number;
  locationY: number;
  startDate: string; // 공연 시작일
  endDate: string; // 공연 종료일
  reservationStartDate: string; // 예매 시작일
  reservationEndDate: string; // 예매 종료일
  price: string;
  rating: number;
  limitAge: number;
  durationTime: number;
  adminId: number;
  concertHallName?: string | null;

  // 조인된 정보
  admin?: Admin;

  // 관계 데이터
  images?: any[];
  seatMap?: any;
  schedules?: any[];
  casts?: any[];
}